requirements.txt
# Smartchime v2.0.0 Requirements
paho-mqtt==1.6.1
# pinned for compatibility with luma.oled 3.14.0
pillow==10.4.0
luma.oled==3.14.0
python-vlc==3.0.21203
gpiozero==2.0
rpi-lgpio==0.6
PyYAML==6.0.1
vcgencmd==0.1.1
pyalsaaudio==0.9.2
fontawesome-free==5.15.4
